fix delete

+++ app/Http/Controllers/SavedRecipeController.php
@@ -20,7 +20,8 @@ class SavedRecipeController extends Controller
 
     public function delete(Request $request){
         $recipeId = $request->recipe_id;
+        $userId = JWTAuth::user()->id;
 
-        SavedRecipe::where('recipe_id', $recipeId)->delete();
+        SavedRecipe::where('user_id', $userId)->where('recipe_id', $recipeId)->first()->delete();
     }
 }
